Krepselio meniu sukurtas

+++ eshop/src/main/java/lt/ju/eshop/Cart/CartItemRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CartItemRepo extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepo extends JpaRepository<CartItem, Long> {
     Optional<CartItem> findByCartClientIdAndProductId(Long cartClientId, Long productId);
 }
